Show jira logout button if the user's connected + it's list of available resources
